+++ config/routes.rb
@@ -20,6 +20,9 @@ JobBoard::Engine.routes.draw do
 
   resources :processes, only: :index
   resources :recurring_tasks, only: :index
+  scope "recurring_tasks/:key", constraints: { key: /[^\/]+/ }, format: false do
+    post "run", to: "recurring_tasks/runs#create", as: :recurring_task_run
+  end
 
   get "assets/:name", to: "assets#show", as: :static_asset, constraints: { name: /[a-z_]+\.(css|js)/ }
 end